[2026-06-18] 권한 분리, mapper id, departmentId 데이터타입 long 변경

+++ src/main/java/com/msa4lms/domain/user/entities/User.java
@@ -6,13 +6,13 @@ import lombok.Setter;
 @Getter
 @Setter
 public class User {
-    private Integer id;
+    private Long id;
     private String loginId;
     private String name;
     private String email;
     private String password;
     private Role role;
-    private Integer departmentId;
+    private Long departmentId;
     private String refreshToken;
     private String createdAt;
     private String updatedAt;
