+++ src/main/java/com/msa4lms/domain/user/responses/UserRes.java
@@ -5,12 +5,12 @@ import lombok.Builder;
 
 @Builder
 public record UserRes(
-        Integer id
+        Long id
         , String loginId
         , String name
         , String email
         , Role role
-        , Integer departmentId
+        , Long departmentId
         , String createdAt
 ) {
 }
